Clear repo-wide lint and format debt so CI gates on green
The python-lint-type-test job was already failing on main (9 pre-existing
ruff findings + format drift in 8 files); this change fixes all of it plus
the 2 findings introduced on this branch, so the job is meaningful again:
move function-level imports to module scope, drop a redundant int() cast,
use an f-string conversion flag, replace ambiguous multiplication signs,
add targeted noqa for the four pre-existing complexity findings, and apply
ruff format. No behavior changes; touched-module tests pass (83).

+++ harness/agent/providers.py
@@ -336,7 +336,7 @@ class AnthropicProvider(LLMProvider):
         uncached = usage.get("input_tokens", 0)
         cache_read = usage.get("cache_read_input_tokens", 0)
         cache_write = usage.get("cache_creation_input_tokens", 0)
-        effective_prompt = int(round(uncached + cache_read * 0.1 + cache_write * 1.25))
+        effective_prompt = round(uncached + cache_read * 0.1 + cache_write * 1.25)
         return LLMResponse(
             content="\n".join(content_text) or None,
             tool_calls=tool_calls,
